sql_chatbot/requirements.txt
streamlit==1.34.0
langchain==0.1.12
langchain-community==0.0.28
langchain_core==0.1.31
openai==1.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pymongo[srv]==4.12.0
pandas==2.2.2
numpy==1.26.4
streamlit-chat==0.1.1 